[BUGFIX] Wrong placement of image caption in text with image
If you omit the first caption, then the following captions are shifted
into this position, meaning the the captions are off by the number of
empty captions in front.

This happens because the captions are stored with a newline "\n" as
separator which is removed by a trim = 1.

Remove this trim from tt_content.image.20.caption.1.1 as the trimming of
the actual content happens with
tt_content.image.20.caption.1.stdWrap.trim = 1.

The change also has a few cleanups that popped up in PhpStorm while
investigation the issue.

Resolves: #48354
Releases: 6.2, 6.1, 6.0
Reviewed-on: https://review.typo3.org/20884
Tested-by: Wouter Wolters
Tested-by: Stefan Neufeind

+++ typo3/sysext/css_styled_content/Classes/Controller/CssStyledContentController.php
@@ -582,7 +582,7 @@ class CssStyledContentController extends \TYPO3\CMS\Frontend\Plugin\AbstractPlug
 				if ($rel) {
 					$imgWidths[$a] = $imgInfo[0] / $rel;
 					// counts the total width of the row with the new height taken into consideration.
-					$relations_cols[floor($a / $colCount)] += $imgWidths[$a];
+					$relations_cols[(int)floor($a / $colCount)] += $imgWidths[$a];
 				}
 			}
 		}
@@ -598,7 +598,7 @@ class CssStyledContentController extends \TYPO3\CMS\Frontend\Plugin\AbstractPlug
 		$rowIdx = 0;
 		for ($a = 0; $a < $imgCount; $a++) {
 			$imgKey = $a + $imgStart;
-			// if the image cannot be interpreted as integer (therefore filename and no FAL id), add the image path
+			// If the image cannot be interpreted as integer (therefore filename and no FAL id), add the image path
 			if (\TYPO3\CMS\Core\Utility\MathUtility::canBeInterpretedAsInteger($imgs[$imgKey])) {
 				$totalImagePath = intval($imgs[$imgKey]);
 			} else {
@@ -730,9 +730,9 @@ class CssStyledContentController extends \TYPO3\CMS\Frontend\Plugin\AbstractPlug
 			// Store the original filepath
 			$origImages[$imgKey] = $GLOBALS['TSFE']->lastImageInfo;
 			if ($GLOBALS['TSFE']->lastImageInfo[0] == 0) {
-				$imageRowsFinalWidths[floor($a / $colCount)] += $this->cObj->data['imagewidth'];
+				$imageRowsFinalWidths[(int)floor($a / $colCount)] += $this->cObj->data['imagewidth'];
 			} else {
-				$imageRowsFinalWidths[floor($a / $colCount)] += $GLOBALS['TSFE']->lastImageInfo[0];
+				$imageRowsFinalWidths[(int)floor($a / $colCount)] += $GLOBALS['TSFE']->lastImageInfo[0];
 			}
 		}
 		// How much space will the image-block occupy?
